requirements.txt
numpy>=1.24.4
opencv-python==4.6.0.66
pillow>=9.5.0
scipy==1.11.1
timm>=0.9.2
tokenizers
torch>=2.0.1
torchaudio>=2.0.2
torchvision>=0.15.2
tqdm>=4.65.0
transformers==4.35.2
xformers>=0.0.20
diffusers==0.25.1
gradio==3.43.1